+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support `entry` `text` bidirectional data-binding
 - Improve examples/basic_entry.rb with bidirectional data-binding for `entry` `text`
 - Improve examples/form.rb with bidirectional data-binding for `entry` `text`
+- Improve examples/form_table.rb with bidirectional data-binding for `entry` `text`
 
 ## 0.4.2